requirements.txt
requests==2.26.0
colorama==0.4.6
openai==0.28.1
rich==10.12.0
python-dotenv==0.19.0
Flask==2.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
flask-cors==4.0.0
numpy>=1.21.0 